7) Feedback - The spelling of verify was wrong in the otp screen Guru - DONE.
8) Feedback -The +91 issue is still not resolved
Guru - DONE.

9) Feedback - When I open a photo there is no way to exit it, like a back button. Also there are some white dots in the top of the screen
Guru - The button is appears for us, need to test it on your device or user can swipe the screen.

11) Feedback - When I click on the create custom message, it takes me to create group, still not sure what is create custom message 
Guru - Removed. That button is for posting text msg to property websites like 99acres, Magicbricks etc… which was pending, because we are waiting for Vikhyat reply on it.